test: apply SCORE_THRESH and float bbon

+++ detectron/utils/vis.py
@@ -604,8 +604,8 @@ def vis_one_image_bbox_2_class(
                 color='white')
 
         if classes[i] == 1 or classes[i] == 2:
-            result_string = str(im_name) + " " + str(score) + " " + str(int(bbox[0])) + " " + str(
-                int(bbox[1])) + " " + str(int(bbox[2] - bbox[0])) + " " + str(int(bbox[3] - bbox[1]))
+            result_string = str(im_name) + " " + str(score) + " " + str(float(bbox[0])) + " " + str(
+                float(bbox[1])) + " " + str(float(bbox[2] - bbox[0])) + " " + str(float(bbox[3] - bbox[1]))
             # print result_string
             result.append(result_string)
